requirements.txt
streamlit
pandas
scikit-learn
twilio
requests
folium
streamlit-folium
polyline
numpy
google-generativeai
